requirements.txt
africastalking==1.2.7
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contextlib2==21.6.0
cryptography==42.0.4
Django==5.0.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.6
jwcrypto==1.5.4
oauthlib==3.2.2
pycparser==2.21
pytz==2024.1
requests==2.31.0
schema==0.7.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1